% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_param.R
\name{gen_param}
\alias{gen_param}
\title{Generate the parameters for the air2water model}
\usage{
gen_param(sim_folder, mean_depth)
}
\arguments{
\item{sim_folder}{filepath; folder which contains the meteorological forcing data}

\item{mean_depth}{numeric; mean depth of the lake (m)}
}
\description{
This calculates the 8 parameters needed to run air2water using the mean depth of
the lake.
}
\examples{
\dontrun{
sim_folder <- system.file('extdata', package = 'air2wateR')
gen_param(sim_folder = sim_folder, mean_depth = 147)
}
}
\author{
Sebastiano Piccoloroaz, Tadhg Moore
}
\keyword{methods}
