% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outputs.R
\name{get_outputs}
\alias{get_outputs}
\title{Extract lake surface temperature}
\usage{
get_outputs(sim_folder)
}
\arguments{
\item{sim_folder}{filepath; the directory where simulation files are contained}
}
\description{
This extracts the modelled temperature and returns a dataframe
}
\examples{
\dontrun{
sim_folder <- system.file('extdata', package = 'air2wateR')
gen_param(sim_folder = sim_folder, mean_depth = 147)
run_air2water(sim_folder = sim_folder)
out <- get_outputs(sim_folder = sim_folder)
out$wml <- out$C * 147 # estimate well-mixed layer as a function of mean depth
}

}
\author{
Sebastiano Piccoloroaz, Tadhg Moore
}
\keyword{methods}
