% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param.R
\name{plot_param}
\alias{plot_param}
\title{Plot the parameters from the calibration of the air2water model}
\usage{
plot_param(sim_folder)
}
\arguments{
\item{sim_folder}{filepath; the directory where simulation files are contained}
}
\description{
This plots the results of the calibration for the model.
}
\examples{
\dontrun{
sim_folder <- system.file('extdata', package = 'air2wateR')
gen_param(sim_folder = sim_folder, mean_depth = 147)
run_air2water(sim_folder = sim_folder)
plot_param(sim_folder = sim_folder)
}
}
\author{
Sebastiano Piccoloroaz, Tadhg Moore
}
\keyword{methods}
