% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_air2water.R
\name{run_air2water}
\alias{run_air2water}
\title{Run the air2water model}
\usage{
run_air2water(sim_folder = ".", mode = "pso", use_pars = TRUE, verbose = FALSE)
}
\arguments{
\item{sim_folder}{filepath; the directory where simulation files are contained}

\item{mode}{character; mode in which the model will be run. There are three options:
\enumerate{
 \item "pso" - Particle swarm operator which runs a calibration routine and then a model simulation with the best parameters (default)
 \item "forward" - runs the model once with a set of provided parameters.
 \item "lhc" - Uses Latin hypercube sampling to calibrate the model and then a model simulation with the best parameters.
}}

\item{use_pars}{boolean; use the parameters from the recent calibration. Only in
"forward" mode. This will use the results from calibration routine in file
"1_XXX_xxx.out" and write to "parameters_forward.txt" before running the model. Defaults to TRUE.}

\item{verbose}{Save output as character vector. Defaults to FALSE}
}
\description{
This runs the air2water model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain valid NML files.
}
\examples{
\dontrun{
sim_folder <- system.file('extdata', package = 'air2wateR')
gen_param(sim_folder = sim_folder, mean_depth = 147)
run_air2water(sim_folder = sim_folder)
}
}
\author{
Tadhg Moore
}
\keyword{methods}
