% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_air2water.R
\name{run_air2water}
\alias{run_air2water}
\title{Run the air2water model}
\usage{
run_air2water(sim_folder = ".", verbose = FALSE)
}
\arguments{
\item{sim_folder}{filepath; the directory where simulation files are contained}

\item{verbose}{Save output as character vector. Defaults to FALSE}

\item{par_file}{filepath; to file with air2water setup. Defaults to 'air2water.par'}
}
\description{
This runs the air2water model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain valid NML files.
}
\examples{
\dontrun{
sim_folder <- system.file('extdata', package = 'air2wateR')
gen_param(sim_folder = sim_folder, mean_depth = 147)
run_air2water(sim_folder = sim_folder)
}
}
\author{
Tadhg Moore
}
\keyword{methods}
